services/workers/requirements.txt
pandas==2.2.2
psycopg2-binary==2.9.9
redis==5.0.7
requests==2.32.3
